+++ apps/evm/test/evm/program_counter_test.exs
@@ -1,0 +1,5 @@
+defmodule EVM.ProgramCounterTest do
+  use ExUnit.Case, async: true
+  doctest EVM.ProgramCounter
+
+end